move data into parley
