+++ .gitattributes
@@ -1,1 +1,1 @@
-unicode_data/src/generated/** linguist-generated
+parley/data/** linguist-generated
