+++ parley/data/icu4x_data/normalizer_nfc_v1.rs.data
