+++ parley/data/icu4x_data/normalizer_nfd_data_v1.rs.data
