+++ parley/data/icu4x_data/normalizer_nfd_supplement_v1.rs.data
