+++ parley/data/icu4x_data/normalizer_nfd_tables_v1.rs.data
