+++ parley/data/icu4x_data/property_enum_bidi_mirroring_glyph_v1.rs.data
