+++ parley/data/icu4x_data/property_name_short_script_v1.rs.data
