+++ parley/data/icu4x_data/segmenter_break_grapheme_cluster_v1.rs.data
