+++ parley/data/icu4x_data/segmenter_break_line_v1.rs.data
