+++ parley/data/icu4x_data/segmenter_break_word_override_v1.rs.data
