+++ parley/data/icu4x_data/segmenter_break_word_v1.rs.data
