+++ parley/data/icu4x_data/segmenter_lstm_auto_v1.rs.data
