+++ parley/src/analysis/cluster.rs
@@ -117,7 +117,7 @@ impl CharCluster {
 
     #[inline(always)]
     fn contributes_to_shaping(ch: char, analysis_data_sources: &AnalysisDataSources) -> bool {
-        let props = analysis_data_sources.composite.properties(ch as u32);
+        let props = analysis_data_sources.properties(ch as u32);
         crate::analysis::contributes_to_shaping(props.general_category(), props.script())
     }
