+++ parley/src/analysis/mod.rs
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 pub(crate) mod cluster;
-mod provider;
+mod data;
 
 use alloc::vec::Vec;
 use core::marker::PhantomData;
 
-use crate::analysis::provider::PROVIDER;
+use crate::analysis::data::{COMPOSITE, Provider};
 use crate::resolve::RangedStyle;
 use crate::{Brush, LayoutContext, WordBreak};
 
@@ -24,7 +24,6 @@ use icu_segmenter::{
     GraphemeClusterSegmenter, GraphemeClusterSegmenterBorrowed, LineSegmenter,
     LineSegmenterBorrowed, WordSegmenter, WordSegmenterBorrowed,
 };
-use unicode_data::CompositeProps;
 
 pub(crate) struct AnalysisDataSources {
     grapheme_segmenter: GraphemeClusterSegmenter,
@@ -34,8 +33,6 @@ pub(crate) struct AnalysisDataSources {
     decomposing_normalizer: CanonicalDecomposition,
     script_short_name: PropertyNamesShort<Script>,
     brackets: CodePointMapData<BidiMirroringGlyph>,
-
-    composite: CompositeProps,
 }
 
 #[derive(Default)]
@@ -62,7 +59,7 @@ impl LineSegmenters {
                     WordBreak::KeepAll => LineBreakWordOption::KeepAll,
                 };
                 line_break_opts.word_option = Some(word_break_strength_icu);
-                LineSegmenter::try_new_auto_unstable(&PROVIDER, line_break_opts)
+                LineSegmenter::try_new_auto_unstable(&Provider, line_break_opts)
                     .expect("Failed to create LineSegmenter")
             })
             .as_borrowed()
@@ -72,24 +69,23 @@ impl LineSegmenters {
 impl AnalysisDataSources {
     pub(crate) fn new() -> Self {
         Self {
-            grapheme_segmenter: GraphemeClusterSegmenter::try_new_unstable(&PROVIDER).unwrap(),
+            grapheme_segmenter: GraphemeClusterSegmenter::try_new_unstable(&Provider).unwrap(),
             word_segmenter: WordSegmenter::try_new_lstm_unstable(
-                &PROVIDER,
+                &Provider,
                 WordBreakOptions::default(),
             )
             .unwrap(),
             line_segmenters: LineSegmenters::default(),
-            composing_normalizer: CanonicalComposition::try_new_unstable(&PROVIDER).unwrap(),
-            decomposing_normalizer: CanonicalDecomposition::try_new_unstable(&PROVIDER).unwrap(),
-            script_short_name: PropertyNamesShort::<Script>::try_new_unstable(&PROVIDER).unwrap(),
-            brackets: CodePointMapData::<BidiMirroringGlyph>::try_new_unstable(&PROVIDER).unwrap(),
-            composite: CompositeProps,
+            composing_normalizer: CanonicalComposition::try_new_unstable(&Provider).unwrap(),
+            decomposing_normalizer: CanonicalDecomposition::try_new_unstable(&Provider).unwrap(),
+            script_short_name: PropertyNamesShort::<Script>::try_new_unstable(&Provider).unwrap(),
+            brackets: CodePointMapData::<BidiMirroringGlyph>::try_new_unstable(&Provider).unwrap(),
         }
     }
 
     #[inline(always)]
-    pub(crate) fn composite(&self) -> &CompositeProps {
-        &self.composite
+    pub(crate) fn properties(&self, ch: u32) -> Properties {
+        Properties(COMPOSITE.get32(ch))
     }
 
     #[inline(always)]
@@ -120,6 +116,153 @@ impl AnalysisDataSources {
     #[inline(always)]
     pub(crate) fn brackets(&self) -> CodePointMapDataBorrowed<'_, BidiMirroringGlyph> {
         self.brackets.as_borrowed()
+    }
+}
+
+/// Unicode character properties relevant for text analysis.
+#[derive(Copy, Clone, Debug)]
+#[allow(unreachable_pub)]
+pub struct Properties(u32);
+
+impl Properties {
+    const SCRIPT_BITS: u32 = 8;
+    const GC_BITS: u32 = 5;
+    const GCB_BITS: u32 = 5;
+    const BIDI_BITS: u32 = 5;
+    const IS_EMOJI_OR_PICTOGRAPH_BITS: u32 = 1;
+    const IS_VARIATION_SELECTOR_BITS: u32 = 1;
+    const IS_REGION_INDICATOR_BITS: u32 = 1;
+    const IS_MANDATORY_LINE_BREAK_BITS: u32 = 1;
+
+    const SCRIPT_SHIFT: u32 = 0;
+    const GC_SHIFT: u32 = Self::SCRIPT_SHIFT + Self::SCRIPT_BITS;
+    const GCB_SHIFT: u32 = Self::GC_SHIFT + Self::GC_BITS;
+    const BIDI_SHIFT: u32 = Self::GCB_SHIFT + Self::GCB_BITS;
+    const IS_EMOJI_OR_PICTOGRAPH_SHIFT: u32 = Self::BIDI_SHIFT + Self::BIDI_BITS;
+    const IS_VARIATION_SELECTOR_SHIFT: u32 =
+        Self::IS_EMOJI_OR_PICTOGRAPH_SHIFT + Self::IS_EMOJI_OR_PICTOGRAPH_BITS;
+    const IS_REGION_INDICATOR_SHIFT: u32 =
+        Self::IS_VARIATION_SELECTOR_SHIFT + Self::IS_VARIATION_SELECTOR_BITS;
+    const IS_MANDATORY_LINE_BREAK_SHIFT: u32 =
+        Self::IS_REGION_INDICATOR_SHIFT + Self::IS_REGION_INDICATOR_BITS;
+
+    /// Packs the given arguments into a single u32.
+    #[cfg(feature = "datagen")]
+    pub fn new(
+        script: Script,
+        gc: GeneralCategory,
+        gcb: GraphemeClusterBreak,
+        bidi: icu_properties::props::BidiClass,
+        is_emoji_or_pictographic: bool,
+        is_variation_selector: bool,
+        is_region_indicator: bool,
+        is_mandatory_linebreak: bool,
+    ) -> Self {
+        let s = script.to_icu4c_value() as u32;
+        let gc = gc as u32;
+        let gcb = gcb.to_icu4c_value() as u32;
+        let bidi = bidi.to_icu4c_value() as u32;
+
+        Self(
+            (s << Self::SCRIPT_SHIFT)
+                | (gc << Self::GC_SHIFT)
+                | (gcb << Self::GCB_SHIFT)
+                | (bidi << Self::BIDI_SHIFT)
+                | ((is_emoji_or_pictographic as u32) << Self::IS_EMOJI_OR_PICTOGRAPH_SHIFT)
+                | ((is_variation_selector as u32) << Self::IS_VARIATION_SELECTOR_SHIFT)
+                | ((is_region_indicator as u32) << Self::IS_REGION_INDICATOR_SHIFT)
+                | ((is_mandatory_linebreak as u32) << Self::IS_MANDATORY_LINE_BREAK_SHIFT),
+        )
+    }
+
+    #[inline(always)]
+    fn get(self, shift: u32, bits: u32) -> u32 {
+        (self.0 >> shift) & ((1 << bits) - 1)
+    }
+
+    /// Returns the script for the character.
+    #[inline(always)]
+    fn script(self) -> Script {
+        #[allow(
+            clippy::cast_possible_truncation,
+            reason = "script data only occupies SCRIPT_BITS bits; we cast to `u16` to fulfil the `from_icu4c_value` contract."
+        )]
+        Script::from_icu4c_value(self.get(Self::SCRIPT_SHIFT, Self::SCRIPT_BITS) as u16)
+    }
+
+    /// Returns the general category for the character.
+    #[inline(always)]
+    fn general_category(self) -> GeneralCategory {
+        #[allow(
+            clippy::cast_possible_truncation,
+            reason = "general category data only occupies GC_BITS bits."
+        )]
+        GeneralCategory::try_from(self.get(Self::GC_SHIFT, Self::GC_BITS) as u8).unwrap()
+    }
+
+    /// Returns the grapheme cluster break for the character.
+    #[inline(always)]
+    fn grapheme_cluster_break(self) -> GraphemeClusterBreak {
+        #[allow(
+            clippy::cast_possible_truncation,
+            reason = "cluster break data only occupies GCB_BITS bits"
+        )]
+        GraphemeClusterBreak::from_icu4c_value(self.get(Self::GCB_SHIFT, Self::GCB_BITS) as u8)
+    }
+
+    /// Returns the bidirectional class for the character.
+    #[inline(always)]
+    fn bidi_class(self) -> icu_properties::props::BidiClass {
+        #[allow(
+            clippy::cast_possible_truncation,
+            reason = "bidi class data only occupies BIDI_BITS bits"
+        )]
+        icu_properties::props::BidiClass::from_icu4c_value(
+            self.get(Self::BIDI_SHIFT, Self::BIDI_BITS) as u8,
+        )
+    }
+
+    /// Returns whether the character is an emoji or pictograph.
+    #[inline(always)]
+    fn is_emoji_or_pictograph(self) -> bool {
+        self.get(
+            Self::IS_EMOJI_OR_PICTOGRAPH_SHIFT,
+            Self::IS_EMOJI_OR_PICTOGRAPH_BITS,
+        ) != 0
+    }
+
+    /// Returns whether the character is a variation selector.
+    #[inline(always)]
+    fn is_variation_selector(self) -> bool {
+        self.get(
+            Self::IS_VARIATION_SELECTOR_SHIFT,
+            Self::IS_VARIATION_SELECTOR_BITS,
+        ) != 0
+    }
+
+    /// Returns whether the character is a region indicator.
+    #[inline(always)]
+    fn is_region_indicator(self) -> bool {
+        self.get(
+            Self::IS_REGION_INDICATOR_SHIFT,
+            Self::IS_REGION_INDICATOR_BITS,
+        ) != 0
+    }
+
+    /// Returns whether the character is a mandatory linebreak.
+    #[inline(always)]
+    fn is_mandatory_linebreak(self) -> bool {
+        self.get(
+            Self::IS_MANDATORY_LINE_BREAK_SHIFT,
+            Self::IS_MANDATORY_LINE_BREAK_BITS,
+        ) != 0
+    }
+}
+
+#[cfg(feature = "datagen")]
+impl From<Properties> for u32 {
+    fn from(value: Properties) -> Self {
+        value.0
     }
 }
 
@@ -445,8 +588,6 @@ pub(crate) fn analyze_text<B: Brush>(lcx: &mut LayoutContext<B>, text: &str) {
         (boundary, ch)
     });
 
-    let composite = lcx.analysis_data_sources.composite();
-
     let mut needs_bidi_resolution = false;
 
     lcx.info.reserve(text.len());
@@ -456,7 +597,7 @@ pub(crate) fn analyze_text<B: Brush>(lcx: &mut LayoutContext<B>, text: &str) {
         // character's index, but we need our iterators to align, and the rest are simply
         // character-indexed.
         .fold(false, |is_mandatory_linebreak, (boundary, ch)| {
-            let properties = composite.properties(ch as u32);
+            let properties = lcx.analysis_data_sources.properties(ch as u32);
             let script = properties.script();
             let grapheme_cluster_break = properties.grapheme_cluster_break();
             let bidi_class = properties.bidi_class();
