--- unicode_data/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "unicode_data"
-keywords = ["text", "layout"]
-categories = ["gui", "graphics"]
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-publish = false
-
-[features]
-default = ["baked"]
-baked = ["dep:zerovec"]
-datagen = []
-
-[dependencies]
-icu_properties = { workspace = true }
-icu_collections = { workspace = true }
-zerovec = { workspace = true, optional = true }
-
-[lints]
-workspace = true
